+++ test/responsive_view_test.dart
@@ -3,14 +3,36 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_clean_architecture/flutter_clean_architecture.dart';
 
 void main() {
+  var stateInitialized = false;
+  var viewDidChangeViewDependenciesTriggered = false;
+  var stateDeactivated = false;
+
+  Widget page;
+  TestController controller;
+
+  setUp(() {
+    controller = TestController(
+      onViewDeactivated: () {
+        stateDeactivated = true;
+      },
+      onViewDidChangeDependencies: () {
+        viewDidChangeViewDependenciesTriggered = true;
+      },
+      onViewInitState: () {
+        stateInitialized = true;
+      },
+    );
+    page = TestPage(controller: controller);
+  });
+
   testWidgets('Run TestPage | Mobile viewport then resizes',
       (WidgetTester tester) async {
-    // Build our app and trigger a frame.
-    final page = TestPage();
-
     await tester.setScreenSize(width: 540, height: 540);
 
     await tester.pumpWidget(MaterialApp(home: page));
+
+    expect(stateInitialized, isTrue);
+    expect(viewDidChangeViewDependenciesTriggered, isTrue);
 
     await tester.pump();
     await tester.pumpAndSettle();
@@ -28,12 +50,14 @@ void main() {
     await tester.pumpAndSettle();
 
     expect(find.text('Desktop'), findsOneWidget);
+
+    // To remove page from tree
+    await tester.pumpWidget(Container());
+
+    expect(stateDeactivated, isTrue);
   });
 
   testWidgets('Run TestPage | Mobile Viewport', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
-    final page = TestPage();
-
     await tester.setScreenSize(width: 540, height: 540);
 
     await tester.pumpWidget(MaterialApp(home: page));
@@ -48,9 +72,6 @@ void main() {
   });
 
   testWidgets('Run TestPage | Tablet Viewport', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
-    final page = TestPage();
-
     await tester.setScreenSize(width: 700, height: 600);
 
     await tester
@@ -63,9 +84,6 @@ void main() {
   });
 
   testWidgets('Run TestPage | Desktop Viewport', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
-    final page = TestPage();
-
     await tester.setScreenSize(width: 1024, height: 1024);
 
     await tester
@@ -78,9 +96,6 @@ void main() {
   });
 
   testWidgets('Run TestPage | Watch Viewport', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
-    final page = TestPage();
-
     await tester.setScreenSize(width: 250, height: 250);
 
     await tester
@@ -94,39 +109,61 @@ void main() {
 }
 
 class TestController extends Controller {
+  final Function onViewDidChangeDependencies;
+  final Function onViewInitState;
+  final Function onViewDeactivated;
+
+  TestController(
+      {this.onViewDidChangeDependencies,
+      this.onViewInitState,
+      this.onViewDeactivated});
+
   @override
   void initListeners() {}
+
+  @override
+  void onInitState() {
+    onViewInitState();
+  }
+
+  @override
+  void onDidChangeDependencies() {
+    onViewDidChangeDependencies();
+  }
+
+  @override
+  void onDeactivated() {
+    onViewDeactivated();
+  }
 }
 
 class TestPage extends View {
-  TestPage({Key key}) : super(key: key);
+  final Controller controller;
+
+  TestPage({Key key, this.controller}) : super(key: key);
 
   @override
-  _TestPageState createState() => _TestPageState(TestController());
+  _TestPageState createState() => _TestPageState(controller);
 }
 
 class _TestPageState extends ResponsiveViewState<TestPage, TestController> {
   _TestPageState(TestController controller) : super(controller);
 
   @override
-  ViewBuilder desktopBuilder = (BuildContext context) {
-    return Container(child: Center(child: Text('Desktop')));
-  };
+  Widget get desktopView =>
+      Container(key: globalKey, child: Center(child: Text('Desktop')));
 
   @override
-  ViewBuilder mobileBuilder = (BuildContext context) {
-    return Container(child: Center(child: Text('Mobile')));
-  };
+  Widget get mobileView =>
+      Container(key: globalKey, child: Center(child: Text('Mobile')));
 
   @override
-  ViewBuilder tabletBuilder = (BuildContext context) {
-    return Container(child: Center(child: Text('Tablet')));
-  };
+  Widget get tabletView =>
+      Container(key: globalKey, child: Center(child: Text('Tablet')));
 
   @override
-  ViewBuilder watchBuilder = (BuildContext context) {
-    return Container(child: Center(child: Text('Watch')));
-  };
+  Widget get watchView =>
+      Container(key: globalKey, child: Center(child: Text('Watch')));
 }
 
 /// This is a snippet to change the default value of test flutter emulator size.